refactor: remove tokenize base
